+++ actions/get_client_id.py
@@ -2,7 +2,7 @@ from lib import action
 
 
 class KeycloakGetClientAction(action.KeycloakBaseAction):
-    def run(self, client_id_name):
+    def run(self, client_name):
 
-        client_id = self.keycloak_admin.get_client_id(client_id_name=client_id_name)
+        client_id = self.keycloak_admin.get_client_id(client_name=client_name)
         return client_id