+++ actions/get_group.py
@@ -1,0 +1,8 @@
+from lib import action
+
+
+class KeycloakgetGroupAction(action.KeycloakBaseAction):
+    def run(self, group_id):
+
+        group = self.keycloak_admin.get_group(group_id=group_id)
+        return group