+++ actions/update_authentication_flow_executions.py
@@ -1,0 +1,7 @@
+from lib import action
+
+
+class KeycloakUpdateAuthenticationFlowExecutionsAction(action.KeycloakBaseAction):
+    def run(self, payload, skip_exists=False):
+
+        self.keycloak_admin.create_authentication_flow(payload=payload, skip_exists=skip_exists)