+++ CHANGELOG.md
@@ -3,3 +3,10 @@
 ## 0.1.0
 
  - Initial Release
+
+## 0.1.3
+
+ - Sync with official python-keycloak lib.
+ - Rework get_group functions.
+ - Add get_idps
+ - Add authentication flow actions